Renaming of session samples
Delete session-fluent as we have session support in hummingbird-auth now
