--- session-persist/Tests/LinuxMain.swift
@@ -1,1 +1,0 @@
-
