+++ session/.dockerignore
