+++ session/Packages/hummingbird-auth
@@ -1,0 +1,1 @@
+/Users/adamfowler/Developer/server/hummingbird-project/hummingbird-auth
